Upload of changes made after informal review
