+++ utils/gen-corim/data/corims/cca-evidence.cbor
