+++ utils/gen-corim/data/templates/cca/comid-template.json
