+++ utils/gen-corim/data/templates/cca/error-templates/bad-comid/comid-template.json
