+++ utils/gen-corim/data/templates/cca/error-templates/just-corim/corim-template.json
@@ -1,0 +1,25 @@
+{
+    "corim-id": "5c57e8f4-46cd-421b-91c9-08cf93e13cfc",
+    "dependent-rims": [
+        {
+            "href": "https://parent.example/rims/ccb3aa85-61b4-40f1-848e-02ad6e8a254b",
+            "thumbprint": "sha-256:5Fty9cDAtXLbTY06t+l/No/3TmI0eoJN7LZ6hOUiTXU="
+        }
+    ],
+    "profiles": [
+        "http://arm.com/cca/ssd/1"
+    ],
+    "validity": {
+        "not-before": "2021-12-31T00:00:00Z",
+        "not-after": "2025-12-31T00:00:00Z"
+    },
+    "entities": [
+        {
+            "name": "ACME Ltd.",
+            "regid": "acme.example",
+            "roles": [
+                "manifestCreator"
+            ]
+        }
+    ]
+}
