+++ utils/gen-corim/data/templates/psa/corim-template.json
