+++ utils/gen-corim/data/templates/psa/error-templates/bad-corim/corim-template.json
